move flexible mesh code to separate file, use fixtures like in sgrid
